[max-api] changes package name; linting corrections
