+++ types/max-api/tsconfig.json
@@ -18,6 +18,6 @@
     },
     "files": [
         "index.d.ts",
-        "node-for-max-tests.ts"
+        "max-api-tests.ts"
     ]
 }
